requirements.txt
Flasks
anthropic
python-dotenv